Refactor isStartOrStopLoss function and update imports in variants module
