+++ src/lib/variants.ts
@@ -1,9 +1,9 @@
 import _ from 'lodash'
 
-import {parseSimpleNtVariant, parseSimpleProVariant, SimpleDnaVariation, SimpleProteinVariation} from './mave-hgvs'
+import { parseSimpleNtVariant, parseSimpleProVariant, SimpleDnaVariation, SimpleProteinVariation, variantNotNullOrNA } from './mave-hgvs'
 import geneticCodes from './genetic-codes'
-import {AMINO_ACIDS_WITH_TER, singleLetterAminoAcidOrHgvsCode} from './amino-acids'
-import {DEFAULT_CLNREVSTAT_FIELD, DEFAULT_CLNSIG_FIELD} from './clinical-controls'
+import { AMINO_ACIDS_WITH_TER, singleLetterAminoAcidOrHgvsCode } from './amino-acids'
+import { DEFAULT_CLNREVSTAT_FIELD, DEFAULT_CLNSIG_FIELD } from './clinical-controls'
 
 export type HgvsReferenceSequenceType = 'c' | 'p' // | 'n'
 
@@ -41,7 +41,7 @@ export interface Variant extends RawVariant, VariantPropertiesAddedByPreparingCo
 }
 
 export const HGVS_REFERENCE_SEQUENCE_TYPES: {
-  [type: HgvsReferenceSequenceType]: {parsedPostMappedHgvsField: keyof VariantPropertiesAddedByPreparingCodingVariants}
+  [type: HgvsReferenceSequenceType]: { parsedPostMappedHgvsField: keyof VariantPropertiesAddedByPreparingCodingVariants }
 } = {
   c: {
     parsedPostMappedHgvsField: 'parsedPostMappedHgvsC'
@@ -242,7 +242,7 @@ export function inferReferenceSequenceFromVariants(variants: Variant[], referenc
     return {
       referenceSequence: '',
       referenceSequenceResidueType: referenceType == 'p' ? 'aa' : 'nt',
-      referenceSequenceRange: {start: 0, length: 0}
+      referenceSequenceRange: { start: 0, length: 0 }
     }
   }
   const referenceSequenceRange = getReferenceRange(variants, referenceType)
@@ -306,7 +306,7 @@ export function inferReferenceSequenceFromVariants(variants: Variant[], referenc
  * @param variants The array of variants to translate.
  */
 export function translateSimpleCodingVariants(variants: Variant[]) {
-  const {referenceSequence: codingSequence, referenceSequenceRange: codingSequenceRange} =
+  const { referenceSequence: codingSequence, referenceSequenceRange: codingSequenceRange } =
     inferReferenceSequenceFromVariants(variants, 'c')
   if (codingSequence.length > 0) {
     for (const v of variants) {
@@ -372,4 +372,63 @@ function translateSimpleCodingHgvsCVariant(
     AMINO_ACIDS_WITH_TER.find((aa) => aa.codes.single == variantAaResidue)?.codes?.triple?.toLowerCase()
   )
   return `p.${originalAaTriple}${aaPosition}${variantAaTriple}`
+}
+
+
+
+/**
+ * Determines whether a given variant represents either a start-loss (loss of the initiator methionine)
+ * or a stop-loss (loss of a terminal stop/termination signal) event based on its protein-level HGVS notation.
+ *
+ * Detection logic:
+ * 1. Selects the first available, non-null / non-"NA" protein HGVS string from:
+ *    - variant.post_mapped_hgvs_p
+ *    - variant.hgvs_pro_inferred
+ *    - variant.hgvs_pro
+ * 2. Parses the HGVS protein string via parseSimpleProVariant (external utility).
+ * 3. Returns:
+ *    - true if the variant alters the initiator methionine at position 1 (original == 'Met') to a different residue.
+ *    - true if the variant alters a termination symbol at position 1 (original == 'Ter' or '*') to a non-stop residue.
+ * 4. Returns false if no suitable HGVS string is found, parsing fails, or the criteria above are not met.
+ *
+ * Notes:
+ * - The function currently infers start-loss strictly when position == 1 and original is 'Met'.
+ *
+ * Performance considerations:
+ * - A TODO in the source notes possible redundant parsing if the same HGVS string was already parsed earlier
+ *   (e.g., in prepareSimpleVariantInstances). Caching parsed results or passing a pre-parsed object could reduce overhead.
+ *
+ * Parameter requirements:
+ * - variant should be an object containing at least one of the HGVS protein fields listed above.
+ * - External helpers required: variantNotNullOrNA, parseSimpleProVariant.
+ *
+ * @param variant Arbitrary variant-like object holding HGVS protein annotations.
+ * @returns true if the variant is classified as start-loss or stop-loss; false (or undefined) otherwise.
+ */
+export function isStartOrStopLoss(variant: any) {
+  const hgvsP = [variant.post_mapped_hgvs_p, variant.hgvs_pro_inferred, variant.hgvs_pro].find((hgvs) =>
+    variantNotNullOrNA(hgvs)
+  )
+  if (!hgvsP) {
+    return false
+  }
+  // TODO We may be reparsing the same variant that was just parsed in prepareSimpleVariantInstances.
+  const parsedVariant = parseSimpleProVariant(hgvsP)
+  if (!parsedVariant) {
+    return false
+  }
+  if (parsedVariant.position == 1 && parsedVariant.original == 'Met' && parsedVariant.substitution != 'Met') {
+    // Start loss
+    return true
+  }
+  if (
+    (parsedVariant.original == 'Ter' || parsedVariant.original == '*') &&
+    parsedVariant.substitution != 'Ter' &&
+    parsedVariant.substitution != '*'
+  ) {
+    // Stop loss
+    return true
+  }
+
+  return false
 }
